Refactoring code smell
